store-api: directly get userAddress from JWT in tokenGate
as opposed to doing this in a prior middleware (fickle, and not making
it clear why this middleware was applied)
