+++ store-api-server/src/server.ts
@@ -2,14 +2,9 @@ import body_parser from 'body-parser';
 const { raw } = body_parser;
 import { NestFactory } from '@nestjs/core';
 import { NestExpressApplication } from '@nestjs/platform-express';
-import { ValidationPipe, Logger } from '@nestjs/common';
+import { ValidationPipe } from '@nestjs/common';
 import { AppModule } from './app.module.js';
-import { BEHIND_PROXY, TOKEN_GATE_SPEC_FILE } from './constants.js';
-import { assertEnv } from './utils.js';
-
-import { createRequire } from 'module';
-const require = createRequire(import.meta.url);
-const jwt = require('jsonwebtoken');
+import { BEHIND_PROXY } from './constants.js';
 
 export async function runKanvasServer() {
   let cors: any = false;
@@ -36,21 +31,6 @@ export function setupKanvasServer(server: NestExpressApplication) {
   server.use('/payment/stripe-webhook', raw({ type: 'application/json' }));
   if (BEHIND_PROXY) {
     server.set('trust proxy', 1);
-  }
-
-  if (typeof TOKEN_GATE_SPEC_FILE !== 'undefined') {
-    const jwtSecret = assertEnv('JWT_SECRET');
-    server.use((req: any, _: any, next: any) => {
-      try {
-        const token = req.get('authorization')?.replace(/^Bearer\ /, '');
-        if (typeof token !== 'undefined') {
-          req.auth = jwt.verify(token, jwtSecret);
-        }
-      } catch (err: any) {
-        Logger.warn(`failed to verify JWT: ${err}`);
-      }
-      next();
-    });
   }
 
   server.enableShutdownHooks();
